requirements_dev_txt.txt
# Development and Testing Dependencies
# Include all production dependencies
-r requirements.txt

# Testing Framework
pytest==7.4.3
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
pytest-flask==1.3.0

# Code Quality and Linting
flake8==6.1.0
black==23.12.0
isort==5.12.0
mypy==1.7.1

# Security Testing
bandit==1.7.5
safety==2.3.4

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Development Server and Debugging
Flask-DebugToolbar==0.14.1

# Environment Management
python-decouple==3.8

# Load Testing
locust==2.17.0

# HTTP Testing
httpx==0.25.2

# Mock and Testing Utilities
responses==0.24.1
factory-boy==3.3.0
faker==20.1.0

# Performance Profiling
py-spy==0.3.14
memory-profiler==0.61.0

# Database Testing (if using databases)
pytest-postgresql==5.0.0

# Continuous Integration Helpers
coverage[toml]==7.3.2
codecov==2.1.13

# API Testing
tavern==2.6.1

# Static Analysis
pylint==3.0.3
prospector==1.10.3

# Type Checking
types-requests==2.31.0.10
types-Flask==1.1.6

# Git Hooks and Pre-commit
pre-commit==3.6.0

# Development Utilities
watchdog==3.0.0
ipython==8.18.1
jupyter==1.0.0

# Dockerfile Testing
hadolint-py==2.12.0.3

# YAML Processing
PyYAML==6.0.1